+++ changelog.md
@@ -17,6 +17,7 @@
     - The drawing now shows the trimmed area with a yellow overlay
     - Displays shortcuts and information in the Sequencer's header
     - Change the trim side on the fly by pressing the letter <kbd>O</kbd>
+    - Toggle ripple edit and select mode on the fly with <kbd>Alt</kbd> and <kbd>Shift</kbd> respectively
 - `Import Local Footage` does not impose a folder structure on you anymore
 
 ### Changes ###